FOUR-12800:Process Actions: Add the tags for new compoments in order to avoid conflicts
