+++ resources/js/processes-catalogue/components/optionsMenu/ChartSaveSearch.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    <!-- TODO -->
+  </div>
+</template>
